backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
google-generativeai==0.4.1
httpx==0.27.0
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

redis==5.0.3
celery==5.3.6
